+++ code/cell.compiler/src/node.bin/cmd/index.ts
@@ -1,0 +1,6 @@
+export * from './cmd.bundle';
+export * from './cmd.clean';
+export * from './cmd.dev';
+export * from './cmd.info';
+export * from './cmd.upload';
+export * from './cmd.watch';
