+++ code/cell.compiler/src/node.bin/types.ts
@@ -1,3 +1,5 @@
+import { minimist } from '../node/common/libs';
+
 export * from '../node/common/types';
 
 export type Commands = Record<string, Command>;
@@ -5,3 +7,5 @@ export type Command = {
   description: string;
   params: Record<string, string>;
 };
+
+export type Argv = minimist.ParsedArgs;
