+++ code/cell.compiler/src/node.bin/util/index.ts
@@ -1,4 +1,5 @@
 export * from './util.hash';
 export * from './util.logger';
 export * from './util.loadConfig';
+export * from './util.param';
 export * from './util.ts';
