+++ code/cell.compiler/src/node.bin/util/util.param.ts
@@ -1,0 +1,8 @@
+import { minimist } from '../common';
+
+type P = minimist.ParsedArgs;
+
+export const nameArg = (argv: P): string | undefined => {
+  const value = argv.name;
+  return Array.isArray(value) ? value[value.length - 1] : value;
+};
